complate layer repository

+++ src/main/java/ir/maktab/data/model/entity/Credit.java
@@ -7,6 +7,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -17,7 +18,9 @@ public class Credit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private double balance;
+
+    private BigDecimal balance;
+
     private String NumberCard;
     private String cvv2;
 }
